Nicolas Regnault: - fix a bug in FQHESphereUnnormalizeState
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1029 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/ParticleOnSphere.h
@@ -238,14 +238,14 @@ class ParticleOnSphere :  public AbstractQHEParticle
   // state = reference to the state to convert
   // reference = set which component has to be normalized to 1
   // return value = converted state
-  virtual RealVector& ConvertToUnnormalizedMonomial(RealVector& state, unsigned int reference = 0);
+  virtual RealVector& ConvertToUnnormalizedMonomial(RealVector& state, long reference = 0);
 
   // convert a state such that its components are now expressed in the normalized basis
   //
   // state = reference to the state to convert
   // reference = set which component has been normalized to 1
   // return value = converted state
-  virtual RealVector& ConvertFromUnnormalizedMonomial(RealVector& state, unsigned int reference = 0);
+  virtual RealVector& ConvertFromUnnormalizedMonomial(RealVector& state, long reference = 0);
 
   // print a given State using the monomial notation
   //
